fix: retain focus when code entry is invalid
Steps to reproduce:

1. Setup a React app that uses the CodeEntry component

```js
import {CodeEntry} from 'cobrowse-agent-ui'

function App() {
    const handleCode(code) {
        // Reject all codes
        return false
    }

    return <CodeEntry onCode={handleCode} />
}
```

2. Enter a 6 digit code

Expected behaviour:

The code entry should shake to indicate the code is invalid. The last
input should remain focused so it is easy to correct the code

Actual behaviour:

The code entry shakes but the last input is not focused so someone has
to click on the input.